Minor change in Homography code
