+++ ip_tests/img_alignment/Image_alignment.py
@@ -1,6 +1,6 @@
 import cv2
 from matplotlib import pyplot as plt
-from Homography import*
+from test_homography_ransac import*
 import numpy as np
 from numpy import linalg as lin
